Switch to Postgres, Also New Tables
Switch to postgres from Sqlite3.  Also create tables
for various items that need to be displayed on forms
(directors, titles, nationalities, etc.)  Also added
a boolean for people that are with CABTAL.

NOTE: You will need to create users in your postgres
database with the correct permissions.  Then you
need to:

    db:drop db:create db:migrate db:seed

for completeness and verify that everything worked
correctly (for development and test databases).