Generate a new request for a new permit request
Use the `sablon` gem to replace mailmerge tags in
docx template (these are stored in
`app/helpers/documents/`).

Since some of the replacement text is dependent
on the user's gender, this commit also enforces
validation on the gender field.

Note: some of the replacement text in the document
(specifically the research notes, future plans,
and the length of the request) are still boilerplate
(filler text), but the idea of the generation is there.

Also, this commit fixes some gem dependency weirdness.