Generate a permit renewal letter.
This adds two helper methods to the person model. First,
a method to find the last letter date (which is the last
time a renewal or permit request letter was sent --
currently the submission date of the most recent "active"
permit is used). Secondly, to find the active research
permit number through a similar method.

This completes the variable substitution and sends the
new document to the user via file download.

There is also error handling built into the model and
controller to prevent someone from attempting to
generate a renew document for a user that has no existing
research permits.  Mainly this is due to the fact that
the renewal letter must referencing existing permit
information which, for that type of user, doesn't exist.

There is still some letter information which is not
currently substituted, namely the gender of the minister,
and the research information for the researcher, and the
request length.  These will be added in future commits.