Languages are spoken in Departments which have Regions.
The research statement for a researcher included in each
letter notes which department each language is spoken in.
Additionally, research can be conducted on multiple
languages under a single request. This commit adds the
relationships to store and retrieve that information and
also a method to build the required string in the correct
format based on the number of languages and the number of
departments in which each is spoken (and according to the
gender of the department).

Also added methods for `future_activities` and
`request_period` which are included on the letters.

All of the letter variables are now being substituted
in each letter except the gender of the minister which
still needs to be added.

Also, adjust tests, adjust models accordingly.

To be done in the next commit:
* Adjust views according to the new model relationships.
* Load data related to the departments